requirements.txt
# Deep Learning Framework
torch>=2.1.0
torchvision>=0.16.0

# Transformers and Model Dependencies
transformers>=4.35.0
accelerate>=0.24.0
deepspeed>=0.12.0

# Data Processing
Pillow>=9.0.0
numpy>=1.21.0
pandas>=1.3.0

# Logging and Monitoring
wandb>=0.16.0
tqdm>=4.64.0
psutil>=5.9.0
pynvml>=11.5.0

# Configuration and Utilities
PyYAML>=6.0.0
scikit-learn>=1.1.0

# Development Tools (optional)
# jupyter>=1.0.0
# matplotlib>=3.5.0
# seaborn>=0.11.0 